# backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a sample server api for beacon based attendance system"
  version: "1.0.0"
  title: "Swagger beacon-based attendance system"
  contact:
    email: "mdsazzadul.islam@tuni.fi"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:8080"
basePath: "/v1"
paths:
...
  /beacon:
    post:
      tags:
      - "beacon"
      summary: "adding a Beacon"
      description: "Add a new Beacon"
      operationId: "addBeacon"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "api_key"
          required: true
          type: "string"
        - in: "body"
          name: "beacon"
          description: "beacon object that needs to be added to the database"
          required: true
          schema:
            $ref: "#/definitions/beaconBody"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid data"
          schema:
            $ref: '#/definitions/responseWithMessage'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Beacon"
    get:
      tags:
      - "beacon"
      summary: "Get all Beacon"
      description: "Get all Beacon from database"
      operationId: "getBeacons"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/beaconsList"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Beacon"
  /beacon/{beaconId}:
    get:
      tags:
      - "beacon"
      summary: "Find an beacon by its ID"
      description: "IDs must be positive integers"
      operationId: "getBeaconById"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - name: "beaconId"
        in: "path"
        description: "ID of beacon that needs to be fetched"
        required: true
        type: "integer"
        minimum: 1
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/beaconItem"
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Beacon"
    put:
      tags:
      - "beacon"
      summary: "Update an beacon by its ID"
      description: "IDs must be positive integers"
      operationId: "updateBeacon"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - name: beaconId
        in: path
        required: true
        type: "integer"
        minimum: 1
        format: "int64"
      - name: "beacon"
        in: "body"
        description: "beacon object that needs to be updated to the database"
        required: true
        schema:
          $ref: "#/definitions/updateBeaconItem"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Beacon"
    delete:
      tags:
      - "beacon"
      summary: "Delete a beacon"
      description: "Delete a beacon from database for path beaconId"
      operationId: "deleteBeacon"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "beaconId"
        required: true
        type: "integer"
        minimum: 1
        format: "int64"
        description: "ID of the beacon"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Beacon"
...
  /course:
    post:
      tags:
      - "course"
      summary: "Add a new course"
      operationId: "addCourse"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Course object in body property"
        required: true
        schema:
          $ref: '#/definitions/addCourse'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid data"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Course"
    get:
      tags:
      - "course"
      summary: "Get all course"
      description: "Get all course from database"
      operationId: "getCourses"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/courseList'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Course"
  /teacherCourses:
    get:
      tags:
      - "course"
      summary: "Get all course where user is course personal"
      description: "Get all course from database where user is course personal"
      operationId: "getTeacherWiseCourses"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/courseList'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Course"
  /course/{courseId}:
    get:
      tags:
      - "course"
      summary: "Find an course by its ID"
      description: "IDs must be positive integers"
      operationId: "getCourseById"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: string
      - name: "courseId"
        in: "path"
        description: "ID of course that needs to be fetched"
        required: true
        type: integer
        minimum: 1
        format: int32
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/courseInfo'
        401:
          description: Authorization failed
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Course"
    put:
      tags:
      - "course"
      summary: "Update an course by its ID"
      description: "ID must be positive integers"
      operationId: "updateCourse"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: path
        name: courseId
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
      - in: "body"
        name: "body"
        description: "course object that needs to be updated to the database"
        required: true
        schema:
          $ref: "#/definitions/updateCourseItem"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid data"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Course"
    delete:
      tags:
      - "course"
      summary: "Delete a course"
      description: "Delete a course from database for path roomId"
      operationId: "deleteCourse"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "courseId"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
        description: "ID of the course"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Course"
...
definitions:
  '404':
    type: object
    additionalProperties: false
    properties:
      error:
        type: string
    required:
      - error
  responseWithMessage:
    type: object
    additionalProperties: false
    properties:
      message:
        type: string
    required:
      - message
...
  beaconBody:
    type: object
    additionalProperties: false
    properties:
      macAddress:
        description: The MAC address of the Bluetooth point. Address with
            colon (:) or hyphen (-) separators, upper or lower case hex digits.
        type: string
        pattern: '^(([a-zA-Z0-9]{2}:){5}[a-zA-Z0-9]{2})|(([a-zA-Z0-9]{2}-){5}[a-zA-Z0-9]{2})$'
    required:
      - macAddress
    example:
      {
        macAddress: '7C:D1:C3:19:BC:EE'
      }
  beaconItem:
    type: object
    additionalProperties: false
    properties:
      id:
        type: integer
        format: int64
      mac_address:
        description: The MAC address of the Bluetooth point. Address with
            colon (:) or hyphen (-) separators, upper or lower case hex digits.
        type: string
        pattern: '^(([a-zA-Z0-9]{2}:){5}[a-zA-Z0-9]{2})|(([a-zA-Z0-9]{2}-){5}[a-zA-Z0-9]{2})$'
      active_status:
        type: boolean
    required:
      - id
      - mac_address
      - active_status
  beaconsList:
    type: array
    items:
      $ref: '#/definitions/beaconItem'
  updateBeaconItem:
    type: object
    additionalProperties: false
    properties:
      macAddress:
        description: The MAC address of the Bluetooth point. Address with
            colon (:) or hyphen (-) separators, upper or lower case hex digits.
        type: string
        pattern: '^(([a-zA-Z0-9]{2}:){5}[a-zA-Z0-9]{2})|(([a-zA-Z0-9]{2}-){5}[a-zA-Z0-9]{2})$'
      activeStatus:
        type: boolean
    required:
      - macAddress
      - activeStatus
    example: {
      macAddress: '7C:D1:C3:19:BC:EE',
        activeStatus: true
      }
...
  addCourse:
    type: object
    properties:
      courseCode:
        type: string
        minLength: 1
      courseName:
        type: string
        minLength: 1
      coursePersonal:
        type: array
        items:
          type: integer
          format: int32
        minItems: 1
        uniqueItems: true
    additionalProperties: false
    required:
      - courseCode
      - courseName
      - coursePersonal
    example: {
        courseCode: "TIE-20106",
        courseName: "Data Structures and Algorithms",
        coursePersonal: [1, 2]
      }
...
  courseInfo:
    type: object
    properties:
      course_id:
        type: integer
        format: int32
      course_code:
        type: string
      course_name:
        type: string
      course_personal:
        type: array
        items:
          type: object
          properties:
            user_id:
              type: integer
              format: int32
            first_name:
              type: string
            last_name:
              type: string
          required:
          - user_id
          - first_name
          - last_name
    additionalProperties: false
    required:
    - course_id
    - course_code
    - course_name
    - course_personal
    example: 
      {
        "course_id": 43,
        "course_code": "TIE-20106",
        "course_name": "Data Structures and Algorithms",
        "course_personal": [
          {
            "id": 2,
            "first_name": "teacher",
            "last_name": "one"
          }
        ]
      }
  courseList:
    type: array
    items:
      $ref: '#/definitions/courseInfo'
  updateCourseItem:
    type: object
    properties:
      courseCode:
        type: string
        minLength: 1
      courseName:
        type: string
        minLength: 1
      coursePersonal:
        type: array
        items:
          type: integer
          format: int32
        minItems: 1
        uniqueItems: true
    example:
      courseCode: "TIE-20106"
      courseName: "Data Structures and Algorithms"
      coursePersonal: [1, 2]
    additionalProperties: false
    required:
      - courseCode
      - courseName
      - coursePersonal
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "X-API-KEY"
    in: "header"